extract kithe_checksum_signature plugin
